pyproject.toml
[build-system]
requires=["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MuseBCI"
version = "0.1.0"
description = "Simple Muse2 EEG BCI Library"
readme = "README.md"
maintainers = [{name = "Gunn Chun", email = "gunncre@gmail.com"}]
classifiers = [
	"Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]
requires-python = "~=3.11.0"
dependencies = [
    "numpy",
    "scikit-learn",
    "scipy",
    "muselsl",
    "pylsl",
    "pyyaml"
]

[tool.setuptools]
packages = ["muse_bci"]